requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
finmind==1.9.0
flask==3.1.2
fonttools==4.61.0
frozenlist==1.8.0
h11==0.16.0
html5lib==1.1
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.3.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
kiwisolver==1.4.9
loguru==0.7.3
lxml==6.0.2
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pluggy==1.6.0
prettytable==3.17.0
prompt-toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.12.5
pydantic-core==2.41.5
pyecharts==2.0.9
pygments==2.19.2
pyparsing==3.2.5
pysocks==1.7.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
selenium==4.38.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
ta==0.5.25
tomli==2.3.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
werkzeug==3.1.4
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
